+++ packages/e2e-contract-tests/playwright/e2e/ForwardAndMintMulticall.test.ts
@@ -8,9 +8,8 @@ import { shortAddress, calculateAssetId } from '../../src/utils';
 import '../../load.envs.js';
 import { testSetup } from '../utils';
 
+import { MAIN_CONTRACT_ID } from './config';
 import { checkFee, connect, checkAddresses } from './utils';
-
-const { VITE_CONTRACT_ID } = process.env;
 
 test.describe('Forward and Mint Multicall', () => {
   let fuelWalletTestHelper: FuelWalletTestHelper;
@@ -55,7 +54,7 @@ test.describe('Forward and Mint Multicall', () => {
     // test mint asset name is shown
     await hasText(walletNotificationPage, 'Unknown', 0, 5000, true);
     // test mint asset id is shown
-    const assetId = calculateAssetId(VITE_CONTRACT_ID!, BaseAssetId);
+    const assetId = calculateAssetId(MAIN_CONTRACT_ID, BaseAssetId);
     await hasText(walletNotificationPage, shortAddress(assetId));
     // test mint amount is correct
     await hasText(walletNotificationPage, mintAmount);
@@ -68,7 +67,7 @@ test.describe('Forward and Mint Multicall', () => {
       maxFee: fee.add(100),
     });
 
-    const fuelContractId = toBech32(VITE_CONTRACT_ID!);
+    const fuelContractId = toBech32(MAIN_CONTRACT_ID);
     await checkAddresses(
       { address: fuelWallet.address.toAddress(), isContract: false },
       { address: fuelContractId, isContract: true },