+++ packages/e2e-contract-tests/playwright/e2e/ForwardHalfAndExternalMint.test.ts
@@ -8,9 +8,8 @@ import { shortAddress, calculateAssetId } from '../../src/utils';
 import '../../load.envs.js';
 import { testSetup } from '../utils';
 
+import { MAIN_CONTRACT_ID, EXTERNAL_CONTRACT_ID } from './config';
 import { checkFee, connect, checkAddresses } from './utils';
-
-const { VITE_CONTRACT_ID, VITE_EXTERNAL_CONTRACT_ID } = process.env;
 
 test.describe('Forward Half ETH and Mint External Custom Asset', () => {
   let fuelWalletTestHelper: FuelWalletTestHelper;
@@ -43,7 +42,6 @@ test.describe('Forward Half ETH and Mint External Custom Asset', () => {
       'Forward Half And External Mint'
     );
     await forwardHalfAndMintButton.click();
-
     const walletNotificationPage =
       await fuelWalletTestHelper.getWalletPopupPage();
 
@@ -64,7 +62,7 @@ test.describe('Forward Half ETH and Mint External Custom Asset', () => {
     // test mint asset name is shown
     await hasText(walletNotificationPage, 'Unknown', 0, 5000, true);
     // test mint asset id is shown
-    const assetId = calculateAssetId(VITE_EXTERNAL_CONTRACT_ID!, BaseAssetId);
+    const assetId = calculateAssetId(EXTERNAL_CONTRACT_ID, BaseAssetId);
     await hasText(walletNotificationPage, shortAddress(assetId));
     // test mint amount is correct
     await hasText(walletNotificationPage, mintAmount);
@@ -77,7 +75,7 @@ test.describe('Forward Half ETH and Mint External Custom Asset', () => {
       maxFee: fee.add(100),
     });
 
-    const fuelContractId = toBech32(VITE_CONTRACT_ID!);
+    const fuelContractId = toBech32(MAIN_CONTRACT_ID);
     await checkAddresses(
       { address: fuelWallet.address.toAddress(), isContract: false },
       { address: fuelContractId, isContract: true },
@@ -85,7 +83,7 @@ test.describe('Forward Half ETH and Mint External Custom Asset', () => {
     );
     await checkAddresses(
       { address: fuelWallet.address.toAddress(), isContract: false },
-      { address: toBech32(VITE_EXTERNAL_CONTRACT_ID!), isContract: true },
+      { address: toBech32(EXTERNAL_CONTRACT_ID), isContract: true },
       walletNotificationPage,
       1,
       1