+++ packages/e2e-contract-tests/src/contracts/contracts/CustomAssetAbi.d.ts
@@ -4,20 +4,22 @@
 /* eslint-disable */
 
 /*
-  Fuels version: 0.63.0
-  Forc version: 0.44.0
-  Fuel-Core version: 0.20.7
+  Fuels version: 0.70.0
+  Forc version: 0.48.1
+  Fuel-Core version: 0.22.0
 */
 
 import type {
   BigNumberish,
   BN,
+  Bytes,
   BytesLike,
   Contract,
   DecodedValue,
   FunctionFragment,
   Interface,
   InvokeFunction,
+  StdString,
 } from 'fuels';
 
 import type { Option, Enum } from './common';